Calculation work
Сontext in English
------------------------------------------------------------------------------------------------------------------------------------------------------------------------
This is a multivariate type of dataset which means providing or involving a variety of separate mathematical or statistical variables, multivariate numerical data analysis. It is composed of 14 attributes which are age, sex, chest pain type, resting blood pressure, serum cholesterol, fasting blood sugar, resting electrocardiographic results, maximum heart rate achieved, exercise-induced angina, oldpeak — ST depression induced by exercise relative to rest, the slope of the peak exercise ST segment, number of major vessels and Thalassemia. This database includes 76 attributes, but all published studies relate to the use of a subset of 14 of them. The Cleveland database is the only one used by ML researchers to date. One of the major tasks on this dataset is to predict based on the given attributes of a patient that whether that particular person has heart disease or not and other is the experimental task to diagnose and find out various insights from this dataset which could help in understanding the problem more.
------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Контекст українською мовою
------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Це багатовимірний тип даних, що передбачає надання або залучення різноманітних окремих математичних чи статистичних змінних для багатовимірного числового аналізу даних. Він складається з 14 атрибутів: вік, стать, тип болю в грудях, рівень артеріального тиску в спокої, рівень холестерину в сироватці крові, рівень цукру в крові натще, результати рестингової електрокардіограми, максимальна досягнута частота серцебиття, стенокардія, викликана фізичним навантаженням, oldpeak — ST-депресія, індукована фізичним навантаженням у порівнянні зі спокоєм, нахил пікового сегмента ST під час вправ, кількість великих судин та таласемія. Ця база даних включає 76 атрибутів, але всі опубліковані дослідження стосуються використання підмножини 14 з них. База даних Клівленда є єдиною, яку до цього часу використовують дослідники машинного навчання. Одне з основних завдань на цих даних — прогнозувати, чи має пацієнт зазначені атрибути серцеві захворювання чи ні, а інше — експериментальне завдання діагностувати та виявити різні інсайти з цього набору даних, які могли б допомогти у більш глибокому розумінні проблеми.